backend/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0

# Embedding models
google-generativeai>=0.3.0

# Clustering
hdbscan>=0.8.33

# Vector operations
faiss-cpu>=1.7.4

# API
fastapi
uvicorn
